changed int ot double

+++ src/org/jfrog/CalculatorTest.java
@@ -9,23 +9,24 @@ import junit.framework.*;
 public class CalculatorTest extends TestCase {
     private Calculator c = new Calculator();
 
-    public void testAdd(){
-        int i = c.calculate(2, '+', 2);
-        assertEquals(4, i);
+    public void testAdd() {
+        double i = c.calculate(2, '+', 2);
+        assertEquals(4.0, i);
     }
 
-    public void testSubtract(){
-        int i = c.calculate(2, '-', 2);
-        assertEquals(0, i);
+    public void testSubtract() {
+        double i = c.calculate(2, '-', 2);
+        assertEquals(0.0, i);
     }
 
-    public void testMultiply(){
-        int i = c.calculate(3, '*', 2);
-        assertEquals(6, i);
+    public void testMultiply() {
+        double i = c.calculate(3, '*', 2);
+        assertEquals(6.0, i);
     }
 
-    public void testDivide(){
-        int i = c.calculate(4, '/', 2);
-        assertEquals(2, i);
+    public void testDivide() {
+        double i = c.calculate(4, '/', 2);
+        assertEquals(2.0, i);
     }
+
 }
